Perform catch-up at block replicator
When a node lags behind the cluster more than the last checkpoint of the leader,
the leader will send a snapshot to it. A snapshot is a block with some raft information.
A received snapshot serves as a trigger for the node to perform catch-up, or state transfer.
It will contact one of the active members of the cluster (preferably the leader), and will
request the missing blocks up to the block indicated by the snapshot.

Signed-off-by: Yoav Tock <tock@il.ibm.com>